initial effective ui
Added updateUI function which calls itself every second and updates the display to show location and capacity of rover and location of drone. Can be updated easily for battery when that's implemented.
Current issue, Rover makes position always end of path so will only show end of path as current location.